Add copy-from-basic script
