+++ templates/experiments/grammaticality-judgment/template-sync/compare-to-basic.test.js
@@ -1,9 +1,7 @@
 // Compares all template files to the basic template
 // Tests will fail is the differences are not permitted in template-diffs.js
 
-const { fs, path } = require("zx");
-const ignore = require("ignore");
-const { diffLines } = require("diff");
+const { basicDir, nonBasicDir, gitignoreDir, compareToBasic } = require("./utils.js");
 let { additions, modifications, deletions } = require("./template-diffs.js");
 
 // For development purposes only, limit the comparison to a shared subdirectory
@@ -14,85 +12,11 @@ let subsetDir;
 // Subset additions, modifications, and deletions based on subsetDir (if provided)
 if (subsetDir) {
   additions = additions.filter((file) => file.startsWith(subsetDir));
-  modifications = modifications.filter((file) => file.startsWith(subsetDir));
+  modifications = modifications.filter((mod) => mod.file.startsWith(subsetDir));
   deletions = deletions.filter((file) => file.startsWith(subsetDir));
 }
 
-// Load .gitignore rules
-const gitignorePath = path.resolve(__dirname, "../../../../.gitignore"); // repo root .gitignore
-const gitignoreRules = fs.readFileSync(gitignorePath, "utf8");
-const ig = ignore().add(gitignoreRules);
-
-/**
- * Recursively list all files in an experiment template directory
- * @param {string} dir - directory to list files from
- * @param {string} templateDir - root directory of the template
- * @param {string} gitignoreDir - directory containing the gitignore file (should be repo root)
- * @param {string[]} fileList - list of files to append to
- * @returns {string[]} - list of files in the template directory
- */
-const listFiles = (dir, templateDir, gitignoreDir, fileList = []) => {
-  const files = fs.readdirSync(dir);
-  files.forEach((file) => {
-    const filePath = path.join(dir, file);
-    const relativePath = path.relative(gitignoreDir, filePath);
-    if (fs.statSync(filePath).isDirectory()) {
-      listFiles(filePath, templateDir, gitignoreDir, fileList);
-    } else if (!ig.ignores(relativePath)) {
-      const templateFile = path.relative(templateDir, filePath);
-      fileList.push(templateFile);
-    }
-  });
-  return fileList;
-};
-
-/**
- * Compares the basic experiment template to a non-basic template
- * @param {string} basicDir - directory of the basic template
- * @param {string} nonBasicDir - directory of the non-basic template
- * @param {string} gitignoreDir - directory containing the gitignore file (should be repo root)
- * @param {string} subsetDir - (for development purposes) limit the template comparison to shared subdirectory
- * @returns {object} - object containing arrays of additions, modifications, and deletions
- */
-const compareToBasic = (basicDir, nonBasicDir, gitignoreDir, subsetDir) => {
-  const additions = [];
-  const modifications = [];
-  const deletions = [];
-  const basicFiles = listFiles(basicDir, basicDir, gitignoreDir);
-  const nonBasicFiles = listFiles(nonBasicDir, nonBasicDir, gitignoreDir);
-  let allFiles;
-  if (subsetDir) {
-    allFiles = new Set(
-      [...basicFiles, ...nonBasicFiles].filter((file) => file.startsWith(subsetDir)),
-    );
-  } else {
-    allFiles = new Set([...basicFiles, ...nonBasicFiles]);
-  }
-  allFiles.forEach((file) => {
-    const basicFile = path.resolve(basicDir, file);
-    const nonBasicFile = path.resolve(nonBasicDir, file);
-    if (fs.existsSync(basicFile) && fs.existsSync(nonBasicFile)) {
-      const basicContent = fs.readFileSync(basicFile, "utf8");
-      const nonBasicContent = fs.readFileSync(nonBasicFile, "utf8");
-      const diffs = diffLines(basicContent, nonBasicContent, { newlineIsToken: true });
-      if (diffs.some((diff) => diff.added || diff.removed)) {
-        modifications.push({ file, diffs });
-      }
-    } else if (fs.existsSync(nonBasicFile)) {
-      additions.push(file);
-    } else {
-      deletions.push(file);
-    }
-  });
-  return { additions, modifications, deletions };
-};
-
-const comparison = compareToBasic(
-  path.resolve(__dirname, "../../basic"),
-  path.resolve(__dirname, ".."),
-  path.dirname(gitignorePath),
-  subsetDir,
-);
+const comparison = compareToBasic(basicDir, nonBasicDir, gitignoreDir, subsetDir);
 
 describe("grammaticality-judgment template", () => {
   test("should have all allowed additions", () => {
