+++ templates/experiments/grammaticality-judgment/template-sync/template-diffs.js
@@ -5,6 +5,7 @@ const additions = [
   "template-sync/compare-to-basic.test.js",
   "template-sync/copy-from-basic.js",
   "template-sync/template-diffs.js",
+  "template-sync/utils.js",
   "src/web page/src/consent.js",
   "src/web page/src/debrief.js",
   "src/web page/src/stim.js",
@@ -44,6 +45,7 @@ const modifications = [
       `"description": "A Pushkin experiment template`,
       `"directory": "templates/experiments/`,
       //`version:`, // Uncomment to allow version to differ
+      `"copy-from-basic": "node template-sync/copy-from-basic.js",`,
     ],
   },
   {
